Drop inlinejs
